src/python/requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.0.3
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
msgpack-python==0.4.8
numpy==1.13.1
pika==0.11.0
pyasn1==0.3.4
pyasn1-modules==0.1.4
PyAudio==0.2.11
pycparser==2.18
pyOpenSSL==17.3.0
pysolr==3.6.0
PyYAML==3.12
pyzmq==16.0.2
requests==2.18.4
service-identity==17.0.0
six==1.10.0
Twisted==17.5.0
urllib3==1.22
watson-developer-cloud==0.26.1
websocket-client==0.44.0
zope.interface==4.4.2
